Lift and shift more get store types
