+++ src/store/get/types/relatedDataPropNameToRelationDict.ts.ts
@@ -1,4 +1,4 @@
-import { DataFormatDeclarations } from '../../dataFormat/types'
+import { DataFormatDeclarations } from '../../../dataFormat/types'
 import {
   RelationDeclarations,
   ExtractRelevantRelationsWithOneToOneToOne,
@@ -7,9 +7,8 @@ import {
   ExtractRelevantRelationsWithOneToManyToMany,
   ExtractRelevantRelationsWithManyToManyFieldRef1,
   ExtractRelevantRelationsWithManyToManyFieldRef2,
-  ExtractForeignFormatNameFromRelation,
-  RelationDeclaration,
-} from '../../relations/types'
+} from '../../../relations/types'
+import { RelatedDataPropertyNamesUnion } from './relatedDataDicts'
 import {
   ManyToManyFieldRef1Name,
   ManyToManyFieldRef2Name,
@@ -19,12 +18,6 @@ import {
   OneToOneToOneName,
 } from './relatedDataPropNames'
 
-type GetForeignDataFormatOfRelation<
-  T extends DataFormatDeclarations,
-  K extends RelationDeclaration,
-  L extends T[number]['name'],
-> = Extract<T[number], { name: ExtractForeignFormatNameFromRelation<K, L> }>
-
 type OneToOneFromOneDict<
   T extends DataFormatDeclarations,
   K extends RelationDeclarations<T>,
@@ -32,7 +25,7 @@ type OneToOneFromOneDict<
 > = {
   [TRelation in ExtractRelevantRelationsWithOneToOneFromOne<L, K> as
     OneToOneFromOneName<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 type OneToOneToOneDict<
@@ -42,7 +35,7 @@ type OneToOneToOneDict<
 > = {
   [TRelation in ExtractRelevantRelationsWithOneToOneToOne<L, K> as
     OneToOneToOneName<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 type OneToManyFromOneDict<
@@ -52,7 +45,7 @@ type OneToManyFromOneDict<
 > = {
   [TRelation in ExtractRelevantRelationsWithOneToManyFromOne<L, K> as
     OneToManyFromOneName<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 type OneToManyToManyDict<
@@ -62,7 +55,7 @@ type OneToManyToManyDict<
 > = {
   [TRelation in ExtractRelevantRelationsWithOneToManyToMany<L, K> as
     OneToManyToManyName<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 type ManyToManyFieldRef1Dict<
@@ -72,7 +65,7 @@ type ManyToManyFieldRef1Dict<
 > = {
   [TRelation in ExtractRelevantRelationsWithManyToManyFieldRef1<L, K> as
     ManyToManyFieldRef1Name<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 type ManyToManyFieldRef2Dict<
@@ -82,13 +75,13 @@ type ManyToManyFieldRef2Dict<
 > = {
   [TRelation in ExtractRelevantRelationsWithManyToManyFieldRef2<L, K> as
     ManyToManyFieldRef2Name<T, TRelation>
-  ]: GetForeignDataFormatOfRelation<T, TRelation, L>
+  ]: TRelation
 }
 
 /**
  * The dict that maps a related data property name to the relation
  */
-export type RelatedDataPropertyNameToForeignDataFormatDict<
+export type RelatedDataPropertyNameToRelationDict<
   T extends DataFormatDeclarations,
   K extends RelationDeclarations<T>,
   L extends T[number]['name'],
@@ -98,3 +91,10 @@ export type RelatedDataPropertyNameToForeignDataFormatDict<
   & OneToManyToManyDict<T, K, L>
   & ManyToManyFieldRef1Dict<T, K, L>
   & ManyToManyFieldRef2Dict<T, K, L>
+
+export type GetRelationOfRelatedDataPropertyName<
+  T extends DataFormatDeclarations,
+  K extends RelationDeclarations<T>,
+  L extends T[number]['name'],
+  P extends RelatedDataPropertyNamesUnion<T, K, L>,
+> = RelatedDataPropertyNameToRelationDict<T, K, L>[P]
